Update styles.css
Hello sir, I made a change here, it has less border radius (Rounded) but more transparency and a improved subtitles, you may like it and you may not liked it. I feel like i want to have a compassion on reviving the project and contribute something to make it better. Dearest thanks.

MRCREEP